# swagger.yaml
swagger: "2.0"
info:
  description: "Server Definition for the Governor. "
  version: "1.0.0"
  title: "Governor"
  termsOfService: "http://servec.com/terms/"
  contact:
    email: "ivamsiravi@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "servec.com"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /notifications:
    post:
      tags: 
      - "notification"
      summary: "Create New Notification"
      operationId: "addNotification"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Notification object"
        required: true
        schema:
          $ref: "#/definitions/Notification"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Notification"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        409:
          description: "Conflict"
    get:
      tags:
      - "notification"
      summary: "Retrieve All Notifications"
      operationId: "getNotifications"
      responses:
        200:
          description: OK
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Notification"
  /notifications/{notificationId}:
    put:
      tags:
      - "notification"
      summary: "Update existing Notification"
      parameters:
      - name: "notificationId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Notification"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        404: 
          description: "Not Found"
    delete:
      tags:
      - "notification"
      summary: "Delete existing Notification"
      parameters:
      - name: "notificationId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Notification"
        404: 
          description: "Not Found"
    get:
      tags:
      - "notification"
      summary: "Retrieve Existing Notification"
      parameters:
      - name: "notificationId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Notification"
        404: 
          description: "Not Found"
...
  /pipelines:
    post:
      tags: 
      - "pipeline"
      summary: "Create New Pipeline"
      operationId: "addPipeline"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Pipeline object"
        required: true
        schema:
          $ref: "#/definitions/Pipeline"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Pipeline"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        409:
          description: "Conflict"
    get:
      tags:
      - "pipeline"
      summary: "Retrieve All Pipeline"
      operationId: "getPipelines"
      responses:
        200:
          description: OK
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Pipeline"
  /pipelines/{pipelineId}:
    put:
      tags:
      - "pipeline"
      summary: "Update existing Pipeline"
      parameters:
      - name: "pipelineId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Pipeline object"
        required: true
        schema:
          $ref: "#/definitions/Pipeline"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Pipeline"
        400: 
          description: "Bad Request"
          schema: 
            $ref: "#/definitions/BadRequest"
        404: 
          description: "Not Found"
    delete:
      tags:
      - "pipeline"
      summary: "Delete existing Pipeline"
      parameters:
      - name: "pipelineId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "No Content"
          schema:
            $ref: "#/definitions/Pipeline"
        404: 
          description: "Not Found"
    get:
      tags:
      - "pipeline"
      summary: "Retrieve Existing Pipeline"
      parameters:
      - name: "pipelineId"
        in: "path"
        description: ""
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/Pipeline"
        404: 
          description: "Not Found"
...
definitions:
  Pipeline:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      components:
        type: "array"
        items:
          $ref: "#/definitions/Component"
      enabled:
        type: "boolean"
  Component:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "integer"
        format: "int64"
      type:
        type: "string"
        enum:
        - "plan"
        - "trigger"
        - "notification"
      sequence:
        type: "integer"
        format: "int64"
...
  Notification:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      to:
        type: "string"
      type:
        type: "string"
      enabled:
        type: "boolean"
...
  BadRequest:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "array"
        items:
          type: "string"
...